+++ tests/unit/models/FolderTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace davidhirtz\yii2\media\tests\unit;
+namespace davidhirtz\yii2\media\tests\unit\models;
 
 use Codeception\Test\Unit;
 use davidhirtz\yii2\media\models\Folder;